# [ch] optimize master_commit_red_percent query